@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "dash-extensions"
-version = "1.0.15"
+version = "1.0.16"
 description = "Extensions for Plotly Dash."
 authors = ["emher <emil.h.eriksen@gmail.com>"]
 license = "MIT"
@@ -23,34 +23,37 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4"
-dash = ">=2.15.0"
-more-itertools = "^9.0.0"
+python = ">=3.9,<4"
+dash = ">=2.17.0"
+more-itertools = ">=10.2.0, <11.0.0"
 jsbeautifier = "^1.14.3"
 Flask-Caching = "^2.1.0"
 dash-mantine-components = {version = "^0.14.3", optional = true}
 dataclass-wizard = "^0.22.2"
+ruff = "^0.4.5"
+pydantic = "^2.7.1"
 
 [tool.poetry.extras]
 mantine = ["dash-mantine-components"]
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 dash = {extras = ["dev", "testing"], version = "^2.17.0"}
 pytest-cov = "^4.0.0"
-black = "^24.3.0"
 pandas = ">=1.5.3"
 quart = "^0.18.3"
 starlette = "^0.36.2"
-requests = "^2.32.0"
+requests = "^2.32.2"
 certifi = "^2023.07.22"
 uvicorn = "^0.20.0"
 sse-starlette = "^1.2.1"
 urllib3 = "^1.26.18"
+pre-commit = "^3.7.1"
+mypy = "^1.10.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.black]
+[tool.ruff]
 line-length = 120
-target-version = ["py39"]
+target-version = "py39"
