@@ -1,26 +1,20 @@
 /* eslint-disable import/prefer-default-export */
-import Lottie from "./components/Lottie.react";
-import WebSocket from "./components/WebSocket.react";
 import BeforeAfter from "./components/BeforeAfter.react";
-import Mermaid from "./components/Mermaid.react";
+import CycleBreaker from "./components/CycleBreaker.react";
 import DeferScript from "./components/DeferScript.react";
-import Purify from "./components/Purify.react";
 import EventListener from "./components/EventListener.react";
 import EventSource from "./components/EventSource.react";
-import CycleBreaker from "./components/CycleBreaker.react";
 import Keyboard from "./components/Keyboard.react";
 import Loading from "./components/Loading.react";
+import Lottie from "./components/Lottie.react";
+import Mermaid from "./components/Mermaid.react";
+import Purify from "./components/Purify.react";
+import SSE from "./components/SSE";
+import WebSocket from "./components/WebSocket.react";
 
 export {
-    Lottie,
-    WebSocket,
-    BeforeAfter,
-    Mermaid,
-    DeferScript,
-    Purify,
-    EventListener,
-    EventSource,
-    CycleBreaker,
-    Keyboard,
-    Loading
+    BeforeAfter, CycleBreaker, DeferScript, EventListener,
+    EventSource, Keyboard,
+    Loading, Lottie, Mermaid, Purify, SSE, WebSocket
 };
+
