@@ -1,4 +1,5 @@
 from contextvars import copy_context
+
 from dash._callback_context import context_value
 from dash._utils import AttributeDict
 
@@ -7,7 +8,7 @@ from tests.mock_app import display, update
 
 def test_update_callback():
     output = update(1, 0)
-    assert output == 'button 1: 1 & button 2: 0'
+    assert output == "button 1: 1 & button 2: 0"
 
 
 def test_display_callback():
@@ -17,4 +18,4 @@ def test_display_callback():
 
     ctx = copy_context()
     output = ctx.run(run_callback)
-    assert output == f'You last clicked button with ID btn-1-ctx-example'
+    assert output == "You last clicked button with ID btn-1-ctx-example"
