@@ -9,14 +9,39 @@ from datetime import datetime
 import dash
 import pandas as pd
 import pytest
-import dash_extensions.enrich
 from dash.exceptions import PreventUpdate
+from pydantic import BaseModel
 
-from dash_extensions.enrich import Output, Input, State, CallbackBlueprint, html, DashProxy, NoOutputTransform, Trigger, \
-    TriggerTransform, MultiplexerTransform, PrefixIdTransform, callback, clientside_callback, DashLogger, LogTransform, \
-    BlockingCallbackTransform, dcc, ServersideOutputTransform, Serverside, ALL, CycleBreakerTransform, \
-    CycleBreakerInput, DependencyCollection, DashBlueprint, MATCH, DataclassTransform
-
+import dash_extensions.enrich
+from dash_extensions.enrich import (
+    ALL,
+    MATCH,
+    BaseModelTransform,
+    BlockingCallbackTransform,
+    CallbackBlueprint,
+    CycleBreakerInput,
+    CycleBreakerTransform,
+    DashBlueprint,
+    DashLogger,
+    DashProxy,
+    DataclassTransform,
+    DependencyCollection,
+    Input,
+    LogTransform,
+    MultiplexerTransform,
+    NoOutputTransform,
+    Output,
+    PrefixIdTransform,
+    Serverside,
+    ServersideOutputTransform,
+    State,
+    Trigger,
+    TriggerTransform,
+    callback,
+    clientside_callback,
+    dcc,
+    html,
+)
 
 # region Test utils/stubs
 
@@ -33,7 +58,7 @@ def _css_escape(s):
 
 
 def _css_selector(_id):
-    return f'#{_css_escape(_stringify_id(_id))}'
+    return f"#{_css_escape(_stringify_id(_id))}"
 
 
 def _cssid(**kwargs):
@@ -44,11 +69,7 @@ def _cssid(**kwargs):
 
 def _get_basic_dash_proxy(**kwargs) -> DashProxy:
     app = DashProxy(**kwargs)
-    app.layout = html.Div([
-        html.Button(id="btn"),
-        html.Div(id="log_server"),
-        html.Div(id="log_client")
-    ])
+    app.layout = html.Div([html.Button(id="btn"), html.Div(id="log_server"), html.Div(id="log_client")])
     return app
 
 
@@ -62,14 +83,20 @@ def _bind_basic_callback(app, flex=False):
 
 
 def _bind_basic_callback_flex(app):
-    @app.callback(output=dict(n=Output("log_server", "children")), inputs=dict(n_clicks=Input("btn", "n_clicks")))
+    @app.callback(
+        output=dict(n=Output("log_server", "children")),
+        inputs=dict(n_clicks=Input("btn", "n_clicks")),
+    )
     def update_log(n_clicks):
         return dict(n=n_clicks)
 
 
 def _bind_basic_clientside_callback(app):
-    app.clientside_callback("function(x){return x;}",
-                            Output("log_client", "children"), Input("btn", "n_clicks"))
+    app.clientside_callback(
+        "function(x){return x;}",
+        Output("log_client", "children"),
+        Input("btn", "n_clicks"),
+    )
 
 
 def _basic_dash_proxy_test(dash_duo, app, element_ids=None, btn_id="btn"):
@@ -88,14 +115,31 @@ def _basic_dash_proxy_test(dash_duo, app, element_ids=None, btn_id="btn"):
 
 
 @pytest.mark.parametrize(
-    'tst, flt',
-    [(dict(a=Input("a", "prop"), b=[Input("b", "prop")], c=dict(ca=[Input("ca1", "prop"), Input("ca2", "prop")])),
-      [Input("a", "prop"), Input("b", "prop"), Input("ca1", "prop"), Input("ca2", "prop")]),
-     ([Input("a", "prop"), Input("b", "prop"), Input("c", "prop")],
-      [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")]),
-     ((Input("a", "prop"), Input("b", "prop"), Input("c", "prop")),
-      [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")])
-     ])
+    "tst, flt",
+    [
+        (
+            dict(
+                a=Input("a", "prop"),
+                b=[Input("b", "prop")],
+                c=dict(ca=[Input("ca1", "prop"), Input("ca2", "prop")]),
+            ),
+            [
+                Input("a", "prop"),
+                Input("b", "prop"),
+                Input("ca1", "prop"),
+                Input("ca2", "prop"),
+            ],
+        ),
+        (
+            [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")],
+            [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")],
+        ),
+        (
+            (Input("a", "prop"), Input("b", "prop"), Input("c", "prop")),
+            [Input("a", "prop"), Input("b", "prop"), Input("c", "prop")],
+        ),
+    ],
+)
 def test_dependency_collection(tst, flt):
     dc = DependencyCollection(tst)
     # Test iteration.
@@ -121,10 +165,14 @@ def test_callback_blueprint():
     cbp = CallbackBlueprint(
         [State("s", "prop"), State("s2", "prop")],
         [Output("o", "prop")],
-        [Input("i", "prop")]
+        [Input("i", "prop")],
     )
     assert list(cbp.outputs) == [Output("o", "prop")]
-    assert list(cbp.inputs) == [State("s", "prop"), State("s2", "prop"), Input("i", "prop")]
+    assert list(cbp.inputs) == [
+        State("s", "prop"),
+        State("s2", "prop"),
+        Input("i", "prop"),
+    ]
     # Test mix.
     cbp = CallbackBlueprint(
         [State("s", "prop"), State("s2", "prop")],
@@ -132,36 +180,30 @@ def test_callback_blueprint():
         [Output("o", "prop")],
         State("s3", "prop"),
         [Input("i", "prop")],
-        Output("o2", "prop")
+        Output("o2", "prop"),
     )
     assert list(cbp.outputs) == [Output("o", "prop"), Output("o2", "prop")]
-    assert list(cbp.inputs) == [State("s", "prop"), State("s2", "prop"), Input("i0", "prop"),
-                                State("s3", "prop"), Input("i", "prop")]
+    assert list(cbp.inputs) == [
+        State("s", "prop"),
+        State("s2", "prop"),
+        Input("i0", "prop"),
+        State("s3", "prop"),
+        Input("i", "prop"),
+    ]
     # Test variables.
     my_input = html.Button()
     my_output = html.Div()
-    cbp = CallbackBlueprint(
-        Input(my_input, "n_clicks"),
-        Output(my_output, "children")
-    )
+    cbp = CallbackBlueprint(Input(my_input, "n_clicks"), Output(my_output, "children"))
     assert list(cbp.outputs) == [Output(my_output, "children")]
     assert list(cbp.inputs) == [Input(my_input, "n_clicks")]
     # Test kwargs.
-    cbp = CallbackBlueprint(
-        Input(my_input, "n_clicks"),
-        Output(my_output, "children"),
-        hello="world"
-    )
+    cbp = CallbackBlueprint(Input(my_input, "n_clicks"), Output(my_output, "children"), hello="world")
     assert cbp.kwargs == dict(hello="world")
 
 
 def test_flexible_callback_signature():
     # Test input/output/state as kwargs.
-    cbp = CallbackBlueprint(
-        output=[Output("o", "prop")],
-        inputs=[Input("i", "prop")],
-        hello="world"
-    )
+    cbp = CallbackBlueprint(output=[Output("o", "prop")], inputs=[Input("i", "prop")], hello="world")
     assert list(cbp.inputs) == [Input("i", "prop")]
     assert list(cbp.outputs) == [Output("o", "prop")]
     assert cbp.kwargs == dict(hello="world")
@@ -169,7 +211,7 @@ def test_flexible_callback_signature():
     cbp = CallbackBlueprint(
         output=dict(o=Output("o", "prop"), u=Output("u", "prop")),
         inputs=dict(i=Input("i", "prop"), s=State("s", "prop")),
-        hello="world"
+        hello="world",
     )
     assert list(cbp.inputs) == [Input("i", "prop"), State("s", "prop")]
     assert list(cbp.outputs) == [Output("o", "prop"), Output("u", "prop")]
@@ -177,24 +219,35 @@ def test_flexible_callback_signature():
     # Test complex dict grouping.
     cbp = CallbackBlueprint(
         output=[Output("o", "prop"), Output("u", "prop")],
-        inputs=dict(w=dict(i=Input("i", "prop"), s=State("s", "prop")), z=dict(i=Input("i2", "prop"))),
-        hello="world"
+        inputs=dict(
+            w=dict(i=Input("i", "prop"), s=State("s", "prop")),
+            z=dict(i=Input("i2", "prop")),
+        ),
+        hello="world",
     )
-    assert list(cbp.inputs) == [Input("i", "prop"), State("s", "prop"), Input("i2", "prop")]
+    assert list(cbp.inputs) == [
+        Input("i", "prop"),
+        State("s", "prop"),
+        Input("i2", "prop"),
+    ]
     assert list(cbp.outputs) == [Output("o", "prop"), Output("u", "prop")]
     assert cbp.kwargs == dict(hello="world")
 
 
 def test_flexible_callback_signature_in_app(dash_duo):
     app = DashProxy()
-    app.layout = html.Div([
-        html.Div(id="a", children="a"),
-        html.Div(id="b", children="b"),
-        html.Div(id="log", children=None),
-    ])
+    app.layout = html.Div(
+        [
+            html.Div(id="a", children="a"),
+            html.Div(id="b", children="b"),
+            html.Div(id="log", children=None),
+        ]
+    )
 
-    @app.callback(output=dict(x=Output("log", "children")),
-                  inputs=dict(dict(g=dict(a=Input("a", "children"), b=Input("b", "children")))))
+    @app.callback(
+        output=dict(x=Output("log", "children")),
+        inputs=dict(dict(g=dict(a=Input("a", "children"), b=Input("b", "children")))),
+    )
     def update_x(g):
         return dict(x=f"{g['a']}_{g['b']}")
 
@@ -215,9 +268,11 @@ def test_dash_proxy(dash_duo):
 def test_dash_output_input_state_compatibility(dash_duo):
     app = _get_basic_dash_proxy()
 
-    @app.callback(dash.Output("log_server", "children"),
-                  dash.Input("btn", "n_clicks"),
-                  dash.State("btn", "n_clicks"))
+    @app.callback(
+        dash.Output("log_server", "children"),
+        dash.Input("btn", "n_clicks"),
+        dash.State("btn", "n_clicks"),
+    )
     def update_log(n_clicks, state):
         return n_clicks
 
@@ -225,14 +280,19 @@ def test_dash_output_input_state_compatibility(dash_duo):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Input("btn", "n_clicks")], dict()),
-     ([], dict(inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Input("btn", "n_clicks")], dict()),
+        ([], dict(inputs=dict(n_clicks=Input("btn", "n_clicks")))),
+    ],
+)
 def test_no_output_transform(dash_duo, args, kwargs):
     app = DashProxy()
-    app.layout = html.Div([
-        html.Button(id="btn"),
-    ])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn"),
+        ]
+    )
 
     @app.callback(*args, **kwargs)
     def update(n_clicks):
@@ -252,19 +312,23 @@ def test_no_output_transform(dash_duo, args, kwargs):
 
 def test_trigger_transform(dash_duo):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[TriggerTransform()])
-    app.layout = html.Div([
-        html.Button(id="btn1"),
-        html.Button(id="btn2"),
-        html.Button(id="btn3"),
-        html.Button(id="btn4"),
-        html.Div(id="log"),
-    ])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn1"),
+            html.Button(id="btn2"),
+            html.Button(id="btn3"),
+            html.Button(id="btn4"),
+            html.Div(id="log"),
+        ]
+    )
 
-    @app.callback(Output("log", "children"),
-                  Trigger("btn1", "n_clicks"),
-                  Input("btn2", "n_clicks"),
-                  Trigger("btn3", "n_clicks"),
-                  State("btn4", "n_clicks"))
+    @app.callback(
+        Output("log", "children"),
+        Trigger("btn1", "n_clicks"),
+        Input("btn2", "n_clicks"),
+        Trigger("btn3", "n_clicks"),
+        State("btn4", "n_clicks"),
+    )
     def update(n_clicks2, n_clicks4):
         return f"{str(n_clicks2)}-{str(n_clicks4)}"
 
@@ -288,13 +352,15 @@ def test_trigger_transform(dash_duo):
 
 def test_trigger_transform_clientside(dash_duo):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[TriggerTransform()])
-    app.layout = html.Div([
-        html.Button(id="btn1"),
-        html.Button(id="btn2"),
-        html.Button(id="btn3"),
-        html.Button(id="btn4"),
-        html.Div(id="log"),
-    ])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn1"),
+            html.Button(id="btn2"),
+            html.Button(id="btn3"),
+            html.Button(id="btn4"),
+            html.Div(id="log"),
+        ]
+    )
 
     app.clientside_callback(
         """(nClicks2, nClicks4) => `${nClicks2}-${nClicks4}`""",
@@ -302,7 +368,8 @@ def test_trigger_transform_clientside(dash_duo):
         Trigger("btn1", "n_clicks"),
         Input("btn2", "n_clicks"),
         Trigger("btn3", "n_clicks"),
-        State("btn4", "n_clicks"))
+        State("btn4", "n_clicks"),
+    )
 
     # Check that the app works.
     dash_duo.start_server(app)
@@ -323,17 +390,32 @@ def test_trigger_transform_clientside(dash_duo):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Output("log", "children"), Input("right", "n_clicks")], dict()),
-     ([], dict(output=[Output("log", "children")], inputs=dict(n_clicks=Input("right", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Output("log", "children"), Input("right", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=[Output("log", "children")],
+                inputs=dict(n_clicks=Input("right", "n_clicks")),
+            ),
+        ),
+    ],
+)
 def test_multiplexer_transform(dash_duo, args, kwargs):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[MultiplexerTransform()])
-    app.layout = html.Div([
-        html.Button(id="left"),
-        html.Button(id="right"),
-        html.Div(id="log"),
-    ])
-    app.clientside_callback("function(x){return 'left'}", Output("log", "children"), Input("left", "n_clicks"))
+    app.layout = html.Div(
+        [
+            html.Button(id="left"),
+            html.Button(id="right"),
+            html.Div(id="log"),
+        ]
+    )
+    app.clientside_callback(
+        "function(x){return 'left'}",
+        Output("log", "children"),
+        Input("left", "n_clicks"),
+    )
 
     @app.callback(*args, **kwargs)
     def update_right(n_clicks):
@@ -353,8 +435,10 @@ def test_multiplexer_transform(dash_duo, args, kwargs):
 
 def test_multiplexer_transform_wildcard(dash_duo):
     def make_callback(i):
-        @app.callback(Output({"type": "div", "id": ALL}, "children"),
-                      Input({"type": f"button{i}", "id": ALL}, "n_clicks"))
+        @app.callback(
+            Output({"type": "div", "id": ALL}, "children"),
+            Input({"type": f"button{i}", "id": ALL}, "n_clicks"),
+        )
         def func(n):
             return [f"Hello from group {i}"] * len(n)
 
@@ -362,7 +446,7 @@ def test_multiplexer_transform_wildcard(dash_duo):
         layout = [
             html.Button(f"Button 0 in group {i}", id={"type": f"button{i}", "id": "x"}),
             html.Button(f"Button 1 in group {i}", id={"type": f"button{i}", "id": "y"}),
-            html.Div(f"Div {i}", id={"type": f"div", "id": i}),
+            html.Div(f"Div {i}", id={"type": "div", "id": i}),
         ]
         make_callback(i)
         return layout
@@ -377,7 +461,7 @@ def test_multiplexer_transform_wildcard(dash_duo):
     assert dash_duo.find_element(_cssid(id="1", type="div")).text == "Hello from group 1"
 
 
-@pytest.mark.parametrize('flex', [False, True])
+@pytest.mark.parametrize("flex", [False, True])
 def test_prefix_id_transform(dash_duo, flex):
     app = _get_basic_dash_proxy(transforms=[PrefixIdTransform(prefix="x")])
     _bind_basic_callback(app, flex)
@@ -387,17 +471,35 @@ def test_prefix_id_transform(dash_duo, flex):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Output("log", "children"), Input("right", "n_clicks")], dict()),
-     ([], dict(output=[Output("log", "children")], inputs=dict(n_clicks=Input("right", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Output("log", "children"), Input("right", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=[Output("log", "children")],
+                inputs=dict(n_clicks=Input("right", "n_clicks")),
+            ),
+        ),
+    ],
+)
 def test_multiplexer_and_prefix_transform(dash_duo, args, kwargs):
-    app = DashProxy(prevent_initial_callbacks=True, transforms=[PrefixIdTransform("prefix"), MultiplexerTransform()])
-    app.layout = html.Div([
-        html.Button(id="left"),
-        html.Button(id="right"),
-        html.Div(id="log"),
-    ])
-    app.clientside_callback("function(x){return 'left'}", Output("log", "children"), Input("left", "n_clicks"))
+    app = DashProxy(
+        prevent_initial_callbacks=True,
+        transforms=[PrefixIdTransform("prefix"), MultiplexerTransform()],
+    )
+    app.layout = html.Div(
+        [
+            html.Button(id="left"),
+            html.Button(id="right"),
+            html.Div(id="log"),
+        ]
+    )
+    app.clientside_callback(
+        "function(x){return 'left'}",
+        Output("log", "children"),
+        Input("left", "n_clicks"),
+    )
 
     @app.callback(*args, **kwargs)
     def update_right(n_clicks):
@@ -417,8 +519,11 @@ def test_multiplexer_and_prefix_transform(dash_duo, args, kwargs):
 
 def test_global_blueprint(dash_duo):
     app = _get_basic_dash_proxy()
-    clientside_callback("function(x){return x;}",
-                        Output("log_client", "children"), Input("btn", "n_clicks"))
+    clientside_callback(
+        "function(x){return x;}",
+        Output("log_client", "children"),
+        Input("btn", "n_clicks"),
+    )
 
     @callback(Output("log_server", "children"), Input("btn", "n_clicks"))
     def update_log(n_clicks):
@@ -431,10 +536,19 @@ def test_global_blueprint(dash_duo):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs, port',
-    [([Output("log", "children"), Input("trigger", "n_intervals")], dict(), 4757),
-     ([], dict(output=[Output("log", "children")],
-               inputs=dict(tick=Input("trigger", "n_intervals"))), 4758)])
+    "args, kwargs, port",
+    [
+        ([Output("log", "children"), Input("trigger", "n_intervals")], dict(), 4757),
+        (
+            [],
+            dict(
+                output=[Output("log", "children")],
+                inputs=dict(tick=Input("trigger", "n_intervals")),
+            ),
+            4758,
+        ),
+    ],
+)
 def test_blocking_callback_transform(dash_duo, args, kwargs, port):
     app = DashProxy(transforms=[BlockingCallbackTransform(timeout=5)])
     app.layout = html.Div([html.Div(id="log"), dcc.Interval(id="trigger", interval=500)])
@@ -467,11 +581,20 @@ def test_blocking_callback_transform_final_invocation(dash_duo):
     f.send_keys("c")
     dash_duo.wait_for_text_to_equal("#log", "abc", timeout=5)  # final invocation
 
+
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Output("store", "children"), Input("btn", "n_clicks")], dict()),
-     ([], dict(output=Output("store", "children"),
-               inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Output("store", "children"), Input("btn", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=Output("store", "children"),
+                inputs=dict(n_clicks=Input("btn", "n_clicks")),
+            ),
+        ),
+    ],
+)
 def test_dataclass_transform(dash_duo, args, kwargs):
     @dataclass
     class StateModel:
@@ -479,11 +602,13 @@ def test_dataclass_transform(dash_duo, args, kwargs):
         list_of_values: list[int]
 
     app = DashProxy(prevent_initial_callbacks=True, transforms=[DataclassTransform()])
-    app.layout = html.Div([
-        html.Button(id="btn"),
-        dcc.Store(id="store"),
-        html.Div(id="log"),
-    ])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn"),
+            dcc.Store(id="store"),
+            html.Div(id="log"),
+        ]
+    )
 
     @app.callback(Output("store", "data"), Input("btn", "n_clicks"))
     def update_default(n_clicks: str):
@@ -502,17 +627,70 @@ def test_dataclass_transform(dash_duo, args, kwargs):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Output("store", "children"), Input("btn", "n_clicks")], dict()),
-     ([], dict(output=Output("store", "children"),
-               inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Output("store", "children"), Input("btn", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=Output("store", "children"),
+                inputs=dict(n_clicks=Input("btn", "n_clicks")),
+            ),
+        ),
+    ],
+)
+def test_base_model_transform(dash_duo, args, kwargs):
+    class StateModel(BaseModel):
+        value: datetime
+        list_of_values: list[int]
+
+    app = DashProxy(prevent_initial_callbacks=True, transforms=[BaseModelTransform()])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn"),
+            dcc.Store(id="store"),
+            html.Div(id="log"),
+        ]
+    )
+
+    @app.callback(Output("store", "data"), Input("btn", "n_clicks"))
+    def update_default(n_clicks: str):
+        return StateModel(value=datetime(2000, 1, 1), list_of_values=[1, 2, 3, 4, 5])
+
+    @app.callback(Output("log", "children"), Input("store", "data"))
+    def update_log(state: StateModel):
+        return f"{state.value.isoformat()}: {str(state.list_of_values)}"
+
+    # Check that stuff works. It doesn't using a normal Dash object.
+    dash_duo.start_server(app)
+    assert dash_duo.find_element("#log").text == ""
+    dash_duo.find_element("#btn").click()
+    time.sleep(0.1)  # wait for callback code to execute.
+    assert dash_duo.find_element("#log").text == "2000-01-01T00:00:00: [1, 2, 3, 4, 5]"
+
+
+@pytest.mark.parametrize(
+    "args, kwargs",
+    [
+        ([Output("store", "children"), Input("btn", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=Output("store", "children"),
+                inputs=dict(n_clicks=Input("btn", "n_clicks")),
+            ),
+        ),
+    ],
+)
 def test_serverside_output_transform(dash_duo, args, kwargs):
     app = DashProxy(prevent_initial_callbacks=True, transforms=[ServersideOutputTransform()])
-    app.layout = html.Div([
-        html.Button(id="btn"),
-        html.Div(id="store"),
-        html.Div(id="log"),
-    ])
+    app.layout = html.Div(
+        [
+            html.Button(id="btn"),
+            html.Div(id="store"),
+            html.Div(id="log"),
+        ]
+    )
 
     @app.callback(*args, **kwargs)
     def update_default(n_clicks):
@@ -532,7 +710,7 @@ def test_serverside_output_transform(dash_duo, args, kwargs):
     assert dash_duo.find_element("#log").text == '{"A":{"0":1}}'
 
 
-def test_serverside_output_transform_wildcard(dash_duo):
+def test_serverside_output_transform_wildcard(dash_duo):  # noqa: C901
     def _id(t, i):
         return {"type": t, "index": i}
 
@@ -543,35 +721,34 @@ def test_serverside_output_transform_wildcard(dash_duo):
             html.Div(id=_id("log", i)),
         ]
 
-    app = DashProxy(prevent_initial_callbacks=True, transforms=[ServersideOutputTransform(), MultiplexerTransform()])
+    app = DashProxy(
+        prevent_initial_callbacks=True,
+        transforms=[ServersideOutputTransform(), MultiplexerTransform()],
+    )
     app.layout = html.Div(make_block("1") + make_block("2") + [html.Div(id="log_all"), html.Button(id="btn_all")])
 
-    @app.callback(Output(_id("sso", MATCH), "children"),
-                  Input(_id("btn", MATCH), "n_clicks"))
+    @app.callback(Output(_id("sso", MATCH), "children"), Input(_id("btn", MATCH), "n_clicks"))
     def update_default(n_clicks):
         """
         Populate serverside output ONE at a time, i.e. WRITE to SSO using MATCH.
         """
         return Serverside(pd.DataFrame(columns=["A"], data=[n_clicks]))
 
-    @app.callback(Output(_id("sso", ALL), "children"),
-                  Input("btn_all", "n_clicks"))
+    @app.callback(Output(_id("sso", ALL), "children"), Input("btn_all", "n_clicks"))
     def update_all(n_clicks):
         """
         Populate serverside output ALL at once, i.e. WRITE to SSO using ALL.
         """
         return tuple([Serverside(pd.DataFrame(columns=["B"], data=[n_clicks]))] * 2)
 
-    @app.callback(Output(_id("log", MATCH), "children"),
-                  Input(_id("sso", MATCH), "children"))
+    @app.callback(Output(_id("log", MATCH), "children"), Input(_id("sso", MATCH), "children"))
     def update_log(data):
         """
         Populate log elements ONE at a time, i.e. READ from SSO using MATCH.
         """
         return data.to_json()
 
-    @app.callback(Output("log_all", "children"),
-                  Input(_id("sso", ALL), "children"))
+    @app.callback(Output("log_all", "children"), Input(_id("sso", ALL), "children"))
     def update_log_all(data):
         """
         Populate ALL log elements, i.e. READ from SSO using ALL.
@@ -607,10 +784,18 @@ def test_serverside_output_transform_wildcard(dash_duo):
 
 
 @pytest.mark.parametrize(
-    'args, kwargs',
-    [([Output("log_server", "children"), Input("btn", "n_clicks")], dict()),
-     ([], dict(output=[Output("log_server", "children")],
-               inputs=dict(n_clicks=Input("btn", "n_clicks"))))])
+    "args, kwargs",
+    [
+        ([Output("log_server", "children"), Input("btn", "n_clicks")], dict()),
+        (
+            [],
+            dict(
+                output=[Output("log_server", "children")],
+                inputs=dict(n_clicks=Input("btn", "n_clicks")),
+            ),
+        ),
+    ],
+)
 def test_log_transform(dash_duo, args, kwargs):
     app = _get_basic_dash_proxy(transforms=[LogTransform(try_use_mantine=False)])
 
@@ -628,18 +813,36 @@ def test_log_transform(dash_duo, args, kwargs):
 
 
 @pytest.mark.parametrize(
-    'c_args, c_kwargs, f_args, f_kwargs',
-    [([Output("celsius", "value"), CycleBreakerInput("fahrenheit", "value")], dict(),
-      [Output("fahrenheit", "value"), Input("celsius", "value")], dict()),
-     ([], dict(output=Output("celsius", "value"), inputs=dict(value=CycleBreakerInput("fahrenheit", "value"))),
-      [], dict(output=Output("fahrenheit", "value"), inputs=dict(value=Input("celsius", "value"))))
-     ])
-def test_cycle_breaker_transform(dash_duo, c_args, c_kwargs, f_args, f_kwargs):
+    "c_args, c_kwargs, f_args, f_kwargs",
+    [
+        (
+            [Output("celsius", "value"), CycleBreakerInput("fahrenheit", "value")],
+            dict(),
+            [Output("fahrenheit", "value"), Input("celsius", "value")],
+            dict(),
+        ),
+        (
+            [],
+            dict(
+                output=Output("celsius", "value"),
+                inputs=dict(value=CycleBreakerInput("fahrenheit", "value")),
+            ),
+            [],
+            dict(
+                output=Output("fahrenheit", "value"),
+                inputs=dict(value=Input("celsius", "value")),
+            ),
+        ),
+    ],
+)
+def test_cycle_breaker_transform(dash_duo, c_args, c_kwargs, f_args, f_kwargs):  # noqa: C901
     app = DashProxy(transforms=[CycleBreakerTransform()])
-    app.layout = html.Div([
-        dcc.Input(id="celsius", type="number"),
-        dcc.Input(id="fahrenheit", type="number"),
-    ])
+    app.layout = html.Div(
+        [
+            dcc.Input(id="celsius", type="number"),
+            dcc.Input(id="fahrenheit", type="number"),
+        ]
+    )
 
     def validate_input(value) -> decimal:
         if value is None:
@@ -659,11 +862,7 @@ def test_cycle_breaker_transform(dash_duo, c_args, c_kwargs, f_args, f_kwargs):
 
     dash_duo.start_server(app)
     time.sleep(0.1)
-    logs = [
-        entry
-        for entry in dash_duo.driver.get_log("browser")
-        if entry["timestamp"] > dash_duo._last_ts
-    ]
+    logs = [entry for entry in dash_duo.driver.get_log("browser") if entry["timestamp"] > dash_duo._last_ts]
     assert len(logs) <= int(os.environ.get("TEST_CYCLE_BREAKER_ALLOWED_ERRORS", "0"))
     f = dash_duo.find_element("#fahrenheit")
     f.send_keys("32")
